[travis] fail the build when multiple tags detected on same commit
When the build is tagged more than once, build.sh currently detects only the
first tag. This makes it likely that the system will build artifacts for a build
number that already exists. By failing the build under these conditions we hope
to avoid wasting cycles on building the wrong thing.